Chainging some lookup logic
This should make it a bit faster in terms of lookup as it no longer
looks at redundant monikers after the top-level entry has been found. We
are now having a separate set of framework monikers for dependencies,
which means that now we can have a better fallback logic.

The general process can still be improved (e.g. if I have a 4.5 base
package, I should not do lookups for upper-level dependencies) but that
is likely to come in a later update.